Moved saltise admin views behind new url
