+++ peerinst/templates/admin/peerinst/index.html
@@ -37,41 +37,6 @@
             {% trans 'Manage group assignments' %}
         </a></p>
       </div>
-      <div class="grp-row">
-        <p><a href="{% url 'admin--new-user-approval' %}">
-            {% trans 'Approve new users' %}
-        </a></p>
-      </div>
-      <div class="grp-row">
-        <p><a href="{% url 'admin:peerinst_message_changelist' %}">
-            {% trans 'Manage messages' %}
-        </a></p>
-      </div>
-      <div class="grp-row">
-        <p><a href="{% url 'admin:peerinst_questionflag_changelist' %}">
-            {% trans 'Flagged questions' %}
-        </a></p>
-      </div>
-      <div class="grp-row">
-        <p><a href="{% url 'admin:peerinst_questionflagreason_changelist' %}">
-            {% trans 'Reasons for flagged questions' %}
-        </a></p>
-      </div>
-      <div class="grp-row">
-        <p><a href="{% url 'admin:peerinst_discipline_changelist' %}">
-            {% trans 'Disciplines' %}
-        </a></p>
-      </div>
-      <div class="grp-row">
-        <p><a href="{% url 'admin--flagged-rationales' %}">
-            {% trans 'Flagged rationales' %}
-        </a></p>
-      </div>
-      <div class="grp-row">
-        <p><a href="{% url 'admin:quality_negwordscriterionrules_change' 1 %}">
-            {% trans 'Manage toxic words' %}
-        </a></p>
-      </div>
     </div>
   </div>
 </div>
