+++ peerinst/templates/peerinst/saltise_admin/group_assignment_management.html
@@ -1,0 +1,76 @@
+{% extends "admin/base_site.html" %}
+{% load i18n static grp_tags %}
+
+{% block title %}{% trans 'Assignment results' %}{% endblock %}
+
+{% block extrastyle %}
+<link href="{% static 'peerinst/css/admin.min.css' %}" rel="stylesheet" type="text/css" />
+{% endblock %}
+
+{% block breadcrumbs %}
+<ul>
+  <li><a href="{% url 'admin-index' %}">{% trans 'Home' %}</a></li>
+  <li>{% trans 'Group assignment distribution' %}</li>
+</ul>
+{% endblock %}
+
+{% block content_title %}
+<h1>{% trans 'Group assignment distribution' %}</h1>
+{% endblock %}
+
+{% block content %}
+<div class="g-d-c-fluid">
+  <form action="" method="get">
+    <fieldset class="grp-module">
+      <div class="grp-row">
+        {{ form.as_p }}
+      </div>
+      <div class="grp-row">
+        <input type="submit" value="{% trans 'Run report' %}" />
+      </div>
+    </fieldset>
+  </form>
+</div>
+
+<div class="g-d-c-fluid">
+<form action="" method="post">
+  {% csrf_token %}
+  <fieldset class="grp-module">
+    <div class="grp-row">
+      <p>{% trans 'Students in group:' %} {{ student_list.count }}</p>
+      <p>{% trans 'Student assignments:' %} {{ student_assignment_list.count }}</p>
+      {% if assignments_created %}
+      <p>{% trans 'Assignments created:' %} {{assignments_created}}</p>
+      <p>{% trans 'Emails sent:' %} {{emails_sent}}</p>
+      {% endif %}
+    </div>
+    <div class="grp-row">
+      <input type="hidden" name="group_assignment" value="{{ group_assignment.pk }}">
+      <input type="submit" value="{% trans 'Send all missing' %}" {% if student_list.count == student_assignment_list.count %}disabled{% endif %} />
+    </div>
+  </fieldset>
+</form>
+</div>
+
+<table>
+    <thead>
+      <tr>
+        <th>#</th>
+        <th>{% trans 'Student' %}</th>
+        <th>{% trans 'Email' %}</th>
+        <th>{% trans 'Assignment' %}</th>
+      </tr>
+    </thead>
+    <tbody>
+      {% for row in student_assignment_list %}
+      <tr class="grp-row-{% cycle 'odd' 'even' %}">
+        <td>{{ forloop.counter }}</td>
+        <td>{{ row.student }}</td>
+        <td>{{ row.student.student.email }}</td>
+        <td>{{ row }}</td>
+      </tr>
+      {% endfor %}
+    </tbody>
+</table>
+
+{% endblock %}
