+++ peerinst/templates/peerinst/saltise_admin/index.html
@@ -1,0 +1,56 @@
+{% extends "admin/base_site.html" %}
+{% load i18n %}
+
+{% block title %}{% trans 'myDalite administration' %}{% endblock %}
+
+{% block content %}
+<h1>{% trans 'Dalite NG administration' %}</h1>
+<a href="{% url 'assignment-list' %}">Back to main site</a>
+<div class="g-d-c">
+  <div class="g-d-12">
+    <div class="grp-module">
+      <h2>{% trans 'Management' %}</h2>
+      <div class="grp-row">
+        <p><a href="{% url 'saltise-admin:group-assignment-management' %}">
+            {% trans 'Manage group assignments' %}
+        </a></p>
+      </div>
+      <div class="grp-row">
+        <p><a href="{% url 'saltise-admin:new-user-approval' %}">
+            {% trans 'Approve new users' %}
+        </a></p>
+      </div>
+      <div class="grp-row">
+        <p><a href="{% url 'admin:peerinst_message_changelist' %}">
+            {% trans 'Manage messages' %}
+        </a></p>
+      </div>
+      <div class="grp-row">
+        <p><a href="{% url 'admin:peerinst_questionflag_changelist' %}">
+            {% trans 'Flagged questions' %}
+        </a></p>
+      </div>
+      <div class="grp-row">
+        <p><a href="{% url 'admin:peerinst_questionflagreason_changelist' %}">
+            {% trans 'Reasons for flagged questions' %}
+        </a></p>
+      </div>
+      <div class="grp-row">
+        <p><a href="{% url 'admin:peerinst_discipline_changelist' %}">
+            {% trans 'Disciplines' %}
+        </a></p>
+      </div>
+      <div class="grp-row">
+        <p><a href="{% url 'saltise-admin:flagged-rationales' %}">
+            {% trans 'Flagged rationales' %}
+        </a></p>
+      </div>
+      <div class="grp-row">
+        <p><a href="{% url 'admin:quality_negwordscriterionrules_change' 1 %}">
+            {% trans 'Manage toxic words' %}
+        </a></p>
+      </div>
+    </div>
+  </div>
+</div>
+{% endblock %}
